services/topo-normalizer/requirements.txt
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
